chore(make-enter-clickable): rename file
